+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All Notable changes to `bisarca/graph` will be documented in this file.
 ## [Unreleased]
 
 ### Added
-- Nothing
+- edges, vertices, graph and sets identification using `Bisarca\Graph\Identifier\IdentifierAwareInterface` and `Bisarca\Graph\Identifier\IdentifierAwareTrait`
 
 ### Deprecated
 - Nothing